consolidate a bit
